Split components up, Implemented Animated Background
