+++ README.md
@@ -2,14 +2,30 @@
 
 ## A simple Linktree clone for collating Social Media Links and displaying offers. Live version on https://connect-bush.netlify.app/
 
-TODO:
+### Features
+
+- VueJs Interface
+- CSS handled by Tailwind/PostCSS
+- Background by particles-bg-vue
+- LinkTree integration
+
+## TODO:
 
 - [x] Update NPM packages
-- [ ] Scope what needs to be refactored
-  - [ ] Show fallback if no available links
-  - [ ] Auth for other user to plug in their airtable deets
-  - [ ] Create animated background
-  - [ ] Make social media links actually go somewhere
+- [x] Scope what needs to be refactored
+
+  - [/] Show fallback if no available links
+  - [x] Separate components
+  - [ ] Reestablish Airtable connection
+  - [x] Create animated background
+
 - [ ] Examine tests and re-write if needed
 - [ ] Write proper readme
+  - [ ] Screenshots
+  - [ ] Instructions for Use (upload to node?)
 - [ ] Upload to Render/Vercel
+
+## FUTURE:
+
+- [ ] Auth for other user to plug in their airtable deets
+- [ ] Make social media links actually go somewhere
