+++ package.json
@@ -26,6 +26,7 @@
     "lodash": "^4.17.21",
     "mini-css-extract-plugin": "^2.9.1",
     "npm": "^6.14.18",
+    "particles-bg-vue": "^1.3.5",
     "postcss": "^7.0.39",
     "postcss-loader": "^3.0.0",
     "tailwindcss": "npm:@tailwindcss/postcss7-compat@^2.2.17",
